@@ -6,71 +6,81 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.entity.Player;
 
 import me.winflix.vitalcore.VitalCore;
-import me.winflix.vitalcore.general.database.collections.TribesCollection;
-import me.winflix.vitalcore.general.database.collections.UsersCollection;
-import me.winflix.vitalcore.general.utils.Utils;
-import me.winflix.vitalcore.tribe.files.UserFile;
-import me.winflix.vitalcore.tribe.files.MessagesFile;
-import me.winflix.vitalcore.tribe.files.TribeFile;
+import me.winflix.vitalcore.addons.config.ChestConfig;
+import me.winflix.vitalcore.core.files.MotdConfigFile;
+import me.winflix.vitalcore.core.files.WorldConfigFile;
+import me.winflix.vitalcore.core.managers.LocaleManager;
+import me.winflix.vitalcore.general.database.collections.tribe.TribesDAO;
+import me.winflix.vitalcore.general.database.collections.tribe.UsersDAO;
+import me.winflix.vitalcore.tribes.files.TribeFile;
+import me.winflix.vitalcore.tribes.files.UserFile;
 
 public class FileManager {
-
-    private static final String TRIBES_FOLDER = "tribes";
-    private static final String USERS_FOLDER = "users";
-    private static final String MESSAGES_FILE_NAME = "en";
-    private static final String I18N_FOLDER = "i18n";
     public List<TribeFile> tribeFiles = new ArrayList<>();
     public List<UserFile> usersFiles = new ArrayList<>();
     public FileConfiguration configFile;
-    public MessagesFile messagesFile;
+    private final LocaleManager localeManager;
+    public MotdConfigFile motdConfigFile;
+    public WorldConfigFile worldConfigFile;
+    public ChestConfig chestsConfig;
     public VitalCore plugin;
 
     public FileManager(VitalCore plugin) {
         this.plugin = plugin;
         configFile = plugin.getConfig();
-        setupFolders();
+        this.localeManager = new LocaleManager(plugin);
         setupConfigFile();
-        setupMessagesFiles();
-        setupTribesFiles();
-        setupUsersFiles();
-        setPrefixes();
+        setupMotdFile();
+        setupWorldConfig();
+        setupAddonsFiles();
     }
 
     public void createFolder(String name) {
         File folder = new File(plugin.getDataFolder(), name);
         if (!folder.exists()) {
-            if (!folder.mkdir()) {
+            if (!folder.mkdirs()) {
                 plugin.getLogger().severe("No se pudo crear la carpeta: " + folder.getPath());
             }
         }
     }
 
-    public void setupFolders() {
-        createFolder(TRIBES_FOLDER);
-        createFolder(USERS_FOLDER);
-        createFolder(I18N_FOLDER);
-    }
-
     public void setupTribesFiles() {
-        tribeFiles = TribesCollection.getAllTribes().stream()
-                .map(tribe -> {
-                    return new TribeFile(plugin, tribe.getId(), TRIBES_FOLDER, tribe);
-                })
+        tribeFiles = TribesDAO.getAllTribes().stream()
+                .map(tribe -> new TribeFile(plugin, tribe.getId().toString(), tribe))
                 .collect(Collectors.toList());
     }
 
     public void setupUsersFiles() {
-        usersFiles = UsersCollection.getAllUsers().stream()
-                .map(player -> {
-                    return new UserFile(plugin, player.getId(), USERS_FOLDER, player);
-                })
+        usersFiles = UsersDAO.getAllUsers().stream()
+                .map(player -> new UserFile(plugin, player.getId().toString(), player))
                 .collect(Collectors.toList());
     }
 
-    public void setupMessagesFiles() {
-        messagesFile = new MessagesFile(plugin, MESSAGES_FILE_NAME, I18N_FOLDER);
+    public void setupMotdFile() {
+        motdConfigFile = new MotdConfigFile(plugin);
+    }
+
+    public void setupWorldConfig() {
+        worldConfigFile = new WorldConfigFile(plugin);
+    }
+
+    private void setupAddonsFiles(){
+        chestsConfig = new ChestConfig(plugin);
+    } 
+
+    public ChestConfig getChestsConfig() {
+        return chestsConfig;
+    }
+
+    public FileConfiguration getWorldsConfig() {
+        return worldConfigFile.getConfig();
+    }
+
+    public void saveWorldsConfig() {
+        worldConfigFile.saveConfig();
     }
 
     public void setupConfigFile() {
@@ -79,8 +89,12 @@ public class FileManager {
         plugin.saveConfig();
     }
 
-    public MessagesFile getMessagesFile() {
-        return messagesFile;
+    public MessagesFile getMessagesFile(Player player) {
+        return localeManager.getMessagesFile(player);
+    }
+
+    public MotdConfigFile getMotdFile() {
+        return motdConfigFile;
     }
 
     public FileConfiguration getConfigFile() {
@@ -110,17 +124,10 @@ public class FileManager {
     }
 
     public void reloadAllFiles() {
-        messagesFile.reloadConfig();
+        localeManager.reloadAllLocales();
+        motdConfigFile.reloadConfig();
+        chestsConfig.reloadConfig();
         plugin.reloadConfig();
-        setupUsersFiles();
-        setupTribesFiles();
-        setPrefixes();
     }
 
-    public void setPrefixes() {
-        FileConfiguration messagesConfig = getMessagesFile().getConfig();
-        Utils.ERROR_PREFIX = messagesConfig.getString("prefixes.error");
-        Utils.INFO_PREFIX = messagesConfig.getString("prefixes.info");
-        Utils.SUCCESS_PREFIX = messagesConfig.getString("prefixes.success");
-    }
 }
