@@ -4,21 +4,21 @@ import java.util.function.Consumer;
 
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
+import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.inventory.ItemStack;
 
-import me.winflix.vitalcore.general.models.PlayerMenuUtility;
 import me.winflix.vitalcore.general.utils.Utils;
 
 public abstract class Menu implements InventoryHolder {
     protected Inventory inventory;
-    protected PlayerMenuUtility playerMenuUtility;
+    protected Player owner;
     protected Consumer<Boolean> afterCloseCallback;
 
-    public Menu(PlayerMenuUtility playerMenuUtility) {
-        this.playerMenuUtility = playerMenuUtility;
+    public Menu(Player owner) {
+        this.owner = owner;
     }
 
     protected ItemStack FILLER_GLASS = new ItemStack(Material.GRAY_STAINED_GLASS_PANE);
@@ -38,11 +38,11 @@ public abstract class Menu implements InventoryHolder {
     public void open() {
         inventory = Bukkit.createInventory(this, getSlots(), Utils.useColors(getMenuName()));
         this.setMenuItems();
-        playerMenuUtility.getOwner().openInventory(inventory);
+        owner.openInventory(inventory);
     }
 
     public void close(boolean condition) {
-        playerMenuUtility.getOwner().closeInventory();
+        owner.closeInventory();
         if (afterCloseCallback != null) {
             afterCloseCallback.accept(condition);
         }
